Fix api routes
Implement logic to actually sign up, login, and logout.
Most importantly we hash the users password so their
original password is not openly saved in the database for
security and privacy reasons. We also generate JWT token to
save user session, the token is signed with a secret key
using the HS256, a HMAC with SHA-256 algorithm.